+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+  - Dashboard detail - fix concatenate log duplicates and rename scroll down to go live
+
 ## [2.4.3] - 2019-09-04
 
 ### Add